requirements.txt
telethon
google-genai
python-dotenv
pydantic
python-telegram-bot
apscheduler
